+++ src/app/endpoints/conversations_v2.py
@@ -6,6 +6,7 @@ from fastapi import APIRouter, Depends, HTTPException, Request
 
 from authentication import get_auth_dependency
 from authorization.middleware import authorize
+from cache.cache_entry import CacheEntry
 from configuration import configuration
 from log import get_logger
 from models.api.requests import ConversationUpdateRequest
@@ -24,7 +25,6 @@ from models.api.responses.successful import (
     ConversationsListResponseV2,
     ConversationUpdateResponse,
 )
-from models.cache_entry import CacheEntry
 from models.common import (
     ConversationTurn,
     Message,
