+++ src/app/endpoints/rlsapi_v1.py
@@ -26,6 +26,7 @@ from configuration import configuration
 from constants import ENDPOINT_PATH_INFER
 from log import get_logger
 from metrics import recording
+from models.api.requests.rlsapi import RlsapiV1InferRequest, RlsapiV1SystemInfo
 from models.api.responses.constants import UNAUTHORIZED_OPENAPI_EXAMPLES
 from models.api.responses.error import (
     ForbiddenResponse,
@@ -37,9 +38,11 @@ from models.api.responses.error import (
     UnauthorizedResponse,
     UnprocessableEntityResponse,
 )
+from models.api.responses.successful.rlsapi import (
+    RlsapiV1InferData,
+    RlsapiV1InferResponse,
+)
 from models.config import Action
-from models.rlsapi.requests import RlsapiV1InferRequest, RlsapiV1SystemInfo
-from models.rlsapi.responses import RlsapiV1InferData, RlsapiV1InferResponse
 from observability import InferenceEventData, build_inference_event, send_splunk_event
 from utils.endpoints import check_configuration_loaded
 from utils.query import (
