+++ src/cache/postgres_cache.py
@@ -6,9 +6,9 @@ import psycopg2
 from psycopg2.extensions import AsIs
 
 from cache.cache import Cache
+from cache.cache_entry import CacheEntry
 from cache.cache_error import CacheError
 from log import get_logger
-from models.cache_entry import CacheEntry
 from models.common import ConversationData
 from models.common.turn_summary import (
     ReferencedDocument,
