+++ src/models/README.md
@@ -3,14 +3,8 @@
 ## [__init__.py](__init__.py)
 Pydantic models.
 
-## [cache_entry.py](cache_entry.py)
-Model for conversation history cache entry.
-
 ## [config.py](config.py)
 Model with service configuration.
-
-## [context.py](context.py)
-Context objects for internal operations.
 
 ## [api/](api/)
 Models for API request and response bodies.
@@ -20,6 +14,3 @@ Shared models and types used across endpoints.
 
 ## [database/](database/)
 SQLAlchemy ORM models.
-
-## [rlsapi/](rlsapi/)
-Models for the rlsapi v1 API surface.
