+++ src/models/api/README.md
@@ -3,5 +3,8 @@
 ## [__init__.py](__init__.py)
 Typed HTTP API models (OpenAPI-oriented) for FastAPI routes. Exposes the [`responses`](responses/README.md) subpackage.
 
+## [requests/](requests/)
+REST request body models, including [`requests/rlsapi.py`](requests/rlsapi.py) for the rlsapi v1 surface.
+
 ## [responses/](responses/README.md)
 HTTP response shapes (successful payloads, errors, and OpenAPI description constants).
