+++ src/models/common/responses/README.md
@@ -3,6 +3,6 @@
 ## [responses_api_params.py](responses_api_params.py)
 Request parameter model for Llama Stack responses API calls.
 
-## [responses_context.py](responses_context.py)
-Request-scoped context model for the responses endpoint pipeline.
+## [contexts.py](contexts.py)
+Context models for pipeline and streaming.
