+++ src/models/common/responses/__init__.py
@@ -1,7 +1,7 @@
 """Shared models for the OpenAI-compatible Responses API pipeline."""
 
+from models.common.responses.contexts import ResponsesContext
 from models.common.responses.responses_api_params import ResponsesApiParams
-from models.common.responses.responses_context import ResponsesContext
 from models.common.responses.responses_conversation_context import (
     ResponsesConversationContext,
 )
