--- src/models/rlsapi/README.md
@@ -1,11 +1,0 @@
-# List of source files stored in `src/models/rlsapi` directory
-
-## [__init__.py](__init__.py)
-Pydantic models for rlsapi v1 integration.
-
-## [requests.py](requests.py)
-Models for rlsapi v1 REST API requests.
-
-## [responses.py](responses.py)
-Models for rlsapi v1 REST API responses.
-
