--- src/models/rlsapi/__init__.py
@@ -1,1 +1,0 @@
-"""Pydantic models for rlsapi v1 integration."""
