+++ tests/unit/app/endpoints/test_streaming_query.py
@@ -68,6 +68,7 @@ from models.api.requests import QueryRequest
 from models.api.responses.error import InternalServerErrorResponse
 from models.common.moderation import ShieldModerationPassed
 from models.common.query import Attachment
+from models.common.responses.contexts import ResponseGeneratorContext
 from models.common.responses.responses_api_params import ResponsesApiParams
 from models.common.turn_summary import (
     RAGChunk,
@@ -76,7 +77,6 @@ from models.common.turn_summary import (
     TurnSummary,
 )
 from models.config import Action
-from models.context import ResponseGeneratorContext
 from utils.stream_interrupts import StreamInterruptRegistry
 from utils.token_counter import TokenCounter
